third_party/WebKit/LayoutTests/http/tests/inspector-enabled/reattach-after-editing-styles.html

<html>
<head>
<link rel="stylesheet" href="resources/style.css">
<script src="../inspector/inspector-test.js"></script>
<script src="../inspector/elements-test.js"></script>
<style>
#testDiv {
    font-family: arial;
}
</style>
<script>

function reopenWebInspector(objectId)
{
    window.didReopen = 1;
    testRunner.closeWebInspector();
    setTimeout(function() {
      testRunner.showWebInspector();
      runTest();
    }, 0);
}

function initTest()
{
    delete window.didReopen;
    if (window.testRunner)
        testRunner.waitUntilDone();
    if (window.testRunner)
        testRunner.showWebInspector();
    runTest();
}

async function test() {
  var result = await TestRunner.RuntimeAgent.evaluate('window.didReopen');

  if (result.type !== 'number') {
    TestRunner.addResult('Opening front-end for the first time');
    TestRunner.addResult('Dump initial styles:');
    dump(onStylesDumped);
  } else {
    TestRunner.addResult('Opening front-end second time');
    TestRunner.addResult('Dump styles after inspector was reopened:');
    dump(TestRunner.completeTest);
  }

  function onStylesDumped() {
    treeElement = ElementsTestRunner.getMatchedStylePropertyTreeItem('color');
    treeElement.startEditing();
    treeElement.nameElement.textContent = 'color';
    treeElement.nameElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));

    // Commit editing.
    treeElement.valueElement.textContent = 'green';
    treeElement.valueElement.firstChild.select();
    treeElement.valueElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
    ElementsTestRunner.waitForStyleApplied(onStyleSheetPropertyEdited);
  }

  function onStyleSheetPropertyEdited() {
    treeElement = ElementsTestRunner.getMatchedStylePropertyTreeItem('font-family');
    treeElement.startEditing();
    treeElement.nameElement.textContent = 'font-family';
    treeElement.nameElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));

    // Commit editing.
    treeElement.valueElement.textContent = 'Helvetica';
    treeElement.valueElement.firstChild.select();
    treeElement.valueElement.dispatchEvent(TestRunner.createKeyEvent('Enter'));
    ElementsTestRunner.waitForStyleApplied(onStyleTagPropertyEdited);
  }

  function onStyleTagPropertyEdited() {
    ElementsTestRunner.addNewRule('#testDiv, my-custom-tag', onFirstRuleAdded);
  }

  function onFirstRuleAdded() {
    ElementsTestRunner.addNewRule('#testDiv, another-custom-tag', onSecondRuleAdded);
  }

  function onSecondRuleAdded() {
    TestRunner.domModel.undo().then(onRuleRemoved);
  }

  function onRuleRemoved() {
    ElementsTestRunner.selectNodeAndWaitForStyles('other', step4);
  }

  function step4() {
    TestRunner.addResult('Dump styles after editing:');
    dump(reopenInspector);
  }

  function reopenInspector() {
    TestRunner.flushResults();
    TestRunner.evaluateInPage('reopenWebInspector()');
  }

  function dump(callback) {
    ElementsTestRunner.selectNodeAndWaitForStyles('testDiv', step2);

    function step2() {
      ElementsTestRunner.dumpSelectedElementStyles(true);
      callback();
    }
  }
}
</script>
</head>

<body onload="initTest()">
<p>This test checks that styles edited through inspector are correctly shown upon reattach.</p>
<div id="testDiv"></div>
<div id="other"></div>
</body>
</html>
